Update tdci.f
Will add the rest of the code (which is now working correctly) today or tomorrow. I still need to work on the parsing, but if necessary we can use it as is since I added the info (the one I needed for my part of the code anyways) to the code itself, i.e. it does not need the output to compute since the info is code into the code.